requirements.txt
crewai>=1.6.1
langchain-sambanova>=0.1.6
python-dotenv>=1.2.1
flask>=2.0.0